fixup! move icons over
