+++ elements/link-modal.js
@@ -3,7 +3,7 @@ const Modal = require('base-elements/modal')
 const morph = require('nanomorph')
 const html = require('choo/html')
 const css = require('sheetify')
-const icon = require('./icon')
+const icon = require('./newIcon')
 
 const prefix = css`
   :host {
@@ -151,13 +151,13 @@ function createModal () {
         </h3>
         <label for="dat-link" class="relative mt4 mb4 ${input}">
           <p class="f7 mt0 mb0 tr absolute confirmation ${confirmClass}">
-            ${icon({id: 'check'})}
+            ${icon('check')}
             Link copied to clipboard
           </p>
           <input name="dat-link" type="text" value=${link} class="relative dib pa0 dat-input-input">
-          ${icon({id: 'link'})}
+          ${icon('link')}
           <button class="absolute pointer dat-input-button" title="Copy to Clipboard" aria-label="Copy to Clipboard" onclick=${onclick}>
-            ${icon({id: 'clipboard'})}
+            ${icon('clipboard')}
           </button>
         </label>
         <p class="f7 color-neutral-70">
@@ -167,7 +167,7 @@ function createModal () {
           onclick=${onexit}
           class="absolute pointer pa0 top-0 right-0 h2 w2 bg-transparent tc exit"
           aria-label="Close Modal">
-          ${icon({id: 'cross'})}
+          ${icon('cross')}
         </button>
       </section>
     `
